Add files via upload
Unsure about what data structures to use:
I've created a class called Vertex, and used an ArrayList of Vertex to track the words; and an Array List of Linked Lists of Vertex for the adjacency list. Do you know a better way to do it?

I've built a first draft of a BFS search based on the above data structures, but it's not elegant - not sure if there's a better way to work on it. I've not built the FindLadder method yet.